+++ ec2.tf
@@ -1,7 +1,7 @@
 ## Security group
 resource "aws_security_group" "this" {
   count       = var.create_security_group && var.create_ec2_instance ? 1 : 0
-  name        = "${var.name}-security-group"
+  name        = "${var.name}${var.resource_name_suffix.security_group}"
   vpc_id      = var.vpc_id
   description = var.security_group_description
   tags        = var.tags
@@ -32,14 +32,14 @@ resource "aws_security_group_rule" "egress" {
 # Iam role
 resource "aws_iam_instance_profile" "this" {
   count = var.create_ec2_instance ? 1 : 0
-  name  = "${var.name}-instance-profile"
+  name  = "${var.name}${var.resource_name_suffix.instance_profile}"
   role  = aws_iam_role.cluster_instance[0].name
   tags  = var.tags
 }
 
 resource "aws_iam_role" "cluster_instance" {
   count              = var.create_ec2_instance ? 1 : 0
-  name               = "${var.name}-cluster-instance-role"
+  name               = "${var.name}${var.resource_name_suffix.iam_role}"
   assume_role_policy = data.aws_iam_policy_document.container_instance.json
   tags               = var.tags
 }
@@ -62,7 +62,7 @@ resource "aws_iam_role_policy_attachment" "ssm" {
 # Launch template
 resource "aws_launch_template" "this" {
   count         = var.create_ec2_instance ? 1 : 0
-  name_prefix   = "${var.name}-launch-template"
+  name_prefix   = "${var.name}${var.resource_name_suffix.launch_template}"
   description   = "Launch template for ${var.name} ECS cluster"
   image_id      = var.image_id
   instance_type = var.instance_type
@@ -86,7 +86,7 @@ resource "aws_launch_template" "this" {
   }
 
   dynamic "block_device_mappings" {
-    for_each = var.block_device_mappings
+    for_each = local.processed_block_device_mappings
     content {
       device_name  = block_device_mappings.value.device_name
       no_device    = lookup(block_device_mappings.value, "no_device", null)
@@ -109,9 +109,9 @@ resource "aws_launch_template" "this" {
   }
 
   metadata_options {
-    http_endpoint               = lookup(var.metadata_options, "http_endpoint", "enabled")
+    http_endpoint               = lookup(var.metadata_options, "http_endpoint", var.metadata_options_defaults.http_endpoint)
     http_put_response_hop_limit = lookup(var.metadata_options, "http_put_response_hop_limit", null)
-    http_tokens                 = lookup(var.metadata_options, "http_tokens", "optional")
+    http_tokens                 = lookup(var.metadata_options, "http_tokens", var.metadata_options_defaults.http_tokens)
     http_protocol_ipv6          = lookup(var.metadata_options, "value.http_protocol_ipv6", null)
     instance_metadata_tags      = lookup(var.metadata_options, "value.instance_metadata_tags", null)
   }
@@ -136,7 +136,7 @@ resource "aws_launch_template" "this" {
 # Auto-scaling Group
 resource "aws_autoscaling_group" "container_instance" {
   count = var.create_ec2_instance ? 1 : 0
-  name  = "${var.name}-asg"
+  name  = "${var.name}${var.resource_name_suffix.autoscaling_group}"
 
   launch_template {
     id      = aws_launch_template.this[0].id
@@ -155,9 +155,9 @@ resource "aws_autoscaling_group" "container_instance" {
   dynamic "tag" {
     for_each = var.enable_managed_scaling ? [1] : []
     content {
-      key                 = "AmazonECSManaged"
-      value               = true
-      propagate_at_launch = true
+      key                 = var.ecs_managed_tag_key
+      value               = var.ecs_managed_tag_value
+      propagate_at_launch = var.ecs_managed_tag_propagate_at_launch
     }
   }
 
@@ -167,7 +167,7 @@ resource "aws_autoscaling_group" "container_instance" {
   min_size           = var.min_size
 
   # Enable instance protection when managed scaling is enabled
-  protect_from_scale_in = var.enable_managed_scaling
+  protect_from_scale_in = var.protect_from_scale_in_independent != null ? var.protect_from_scale_in_independent : var.enable_managed_scaling
 
   lifecycle {
     create_before_destroy = true